Fix crash when loading album art by resource
When the album art is loaded from the notification largeIcon it may be a
resource rather than a bitmap, so it has to be loaded as a drawable.
Bitmaps are still loaded using `WallpaperColors.fromBitmap` to avoid
the extra steps needed by the `fromDrawable` method.

Fixes: 239368697
Test: manual
Test: atest MediaControlPanelTest
Change-Id: I0ba9a4b0f4d308cc360a785b68cbabdcb2480690
(cherry picked from commit 29dff3fa8901160518abd40e462ba0d073334318)