Catch failure to set screenshot sound
In some cases, MediaPlayer.create fails and propagates up an
IllegalStateException, which causes the screenshot process to
crash. Catch this error (and just don't play the screenshot sound
if it fails to load).

Bug: 239236526
Fix: 239236526
Test: forced the setDataSource to throw an IOException (which is
silently swallowed by MediaPlayer; see bug; and then thrown
as an IllegalStateException since the data source was never
actually set)

Change-Id: I7ae6ae4cc69b8e3eccfbb7df7558787e50db755f