Add @RequiresNoPermission to IInputMethodManager#startProtoDump()
This is a follow up CL to our previous CL [1], which introduced

  IInputMethodManager#startProtoDump()

so that IME client process can upload its own IME tracing dump data
to InputMethodManagerService.

One may notice that there is no permission / caller verification in

  InputMethodManagerService#startProtoDump(),

but the absence of a permission check there is by design. We need
unprivileged app processes to be able to send that tracing data. To
make it clear in the AIDL file, this CL adds @RequiresNoPermission to
that IPC method.

The annotation is basically for errorprone check [2].  There should be
no observable behavior change at the run time.

 [1]: Ia89f11d5ef8a220ea7746191b18769cea5a8359d
      f44468cb02135a2f8ae8b58b7318b084aec9d904
 [2]: I26a872f07ab13931c241cbb02ff7228edf7dc3b9
      c31b0bec0d8b2febf6fdbb7c776e024b0821c824

Bug: 204904070
Bug: 237316307
Test: presubmit
Change-Id: Ia41be32333a668cbf112ba7394bcca8c8ba7db6a