Inline IMMS#startInputOrWindowGainedFocusInternal
This is a follow up CL to my previous CL [1], which aimed to be a
logical revert of changes mentioned in the commit message.

One thing that CL forgot to handle was

  InputMethodManagerService#startInputOrWindowGainedFocusInternal,

which was mentioned in one of our previous CLs [2].

This CL inlines that method again for simplicity.

This is a mechanical refactoring.  There should be no behavior change.

 [1]: If16ac0de536d9089eb04f6e07b1ee47378124658
      4e3c28ffa0f32f420576d871312b072f4ccaa395
 [2]: Ic584203c1221fbae17f5e2d8f09e3992df061646
      03645899518a3414ebebe4c975eafd4805f777d7

Bug: 234882948
Test: presubmit
Change-Id: I7c775245882815483f29acfd83f53b08a40d07ed