[3/?] Reduce BroadcastQueue interface complexity.
Instead of requesting a specific concrete action of the broadcast
stack, pivot to communicating application lifecycle changes, which
the broadcast stack can then respond to internally.

Slight rename to skipCurrentOrPendingReceiverLocked() to reflect
that it will skip either the current or pending receiver.  This
renamed method is then called via onApplicationTimeoutLocked(),
which previously tried applying pending-vs-current nuance where
it appears to have not been needed.

Bug: 243656033
Test: atest CtsContentTestCases:BroadcastReceiverTest
Change-Id: I451df11e499465940658e15d730491f44add9b62