Handle exception thrown by KCS.setGrant()
KeyChainService.setGrant() now throws an IllegalArguementException if
the alias passed to it does not have an exisitng key. So
DPMS.setKeyGrantForApp() and DMPS.setKeyGrantToWifiAuth() will also now
throw IllegalArgumentExceptions in the senario described above. As this
will propagaet up to the DPM API's that call these functions, we will
have to gate this new behavios on DPC's target SDK level.

So if DPC targets DPC U+ : throw exception
else : return false

Fixes: 175101461
Test: n/a
Topic: make-GrantKeyPairToApp-And-grantKeyPairToWifiAuth-Throw-Exception
Change-Id: I995227b633e2fd60dad096e35f92cb539a9389d4