Remove MODE_WAKE_AND_UNLOCK_PULSING from stopDozing
A race condition exists when unlocking with UDFPS, or when unlocking
from a pulse state, that leaves SystemUI believing it is still in a
doze state even after unlocking. Because of this, the blur never
transitions and instead stuck in the doze state, which then causes
SystemUI to retain focus and grab all touches over the launcher.

This code path should no longer be relevant. Updated unlock animations
now handle all the keyguard elements without the need for this
dedicated path.

Furthermore, cleanup unused paths in
NotifcationShadeDepthController. The animateTo() calls accepted an
optional view which was never used.

Test: manual - Use UDFPS to unlock while dozing with AOD, without AOD,
and when pulsing from a notification
Test: atest NotificationShadeDepthControllerTest
CentralSurfacesImplTest DozeServiceHostTest
Fixes: 242415456

Change-Id: Ide66ab84eda91d8938a48014a91cbe3f12f2bf26
(cherry picked from commit 48b4e69f38a0d0cfb587251791c9dc787bd32cf7)