Initial pass at "Snowbird" broadcast queue.
Now that we have a solid platform carved out, we can start a second
BroadcastQueue implementation with the new ideas we're pursuing for
the "Snowbird" project.

Generally speaking, this new implementation pivots broadcast dispatch
to be handled on a per-process basis, as represented by the new
BroadcastProcessQueue class.  Each per-process queue has the concept
of being "runnable at" a specific time in the future, which supports
us pausing or delaying broadcast delivery on a per-process basis.

This initial pass is just enough of the functionality to pass the
local unit tests we've recently written, which currently cover very
basic situations like manifest-vs-registered and warm-vs-cold starts;
much of the remaining work is left as "TODO" statements, which will
be fleshed out via TDD.

Bug: 245771249
Test: atest FrameworksMockingServicesTests:BroadcastQueueTest
Change-Id: If83377c2012dd16fe8f62f90704263a977763dc7